+++ src/Commands/Pages/RemovePageComponent.cs
@@ -3,11 +3,13 @@ using PnP.PowerShell.Commands.Base.PipeBinds;
 using System;
 using System.Linq;
 using System.Management.Automation;
+using PnP.PowerShell.Commands.Attributes;
 
 namespace PnP.PowerShell.Commands.Pages
 {
     [Cmdlet(VerbsCommon.Remove, "PnPPageComponent")]
     [Alias("Remove-PnPClientSideComponent")]
+    [WriteAliasWarning("Please use 'Remove-PnPPageComponent'. The alias 'Remove-PnPClientSidecComponent' will be removed in the 1.5.0 release")]
     public class RemovePageComponent : PnPWebCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0)]
@@ -27,8 +29,8 @@ namespace PnP.PowerShell.Commands.Pages
                 throw new Exception($"Page '{Page?.Name}' does not exist");
 
             var control = clientSidePage.Controls.FirstOrDefault(c => c.InstanceId == InstanceId);
-            if(control != null)
-            { 
+            if (control != null)
+            {
                 if (Force || ShouldContinue(string.Format(Properties.Resources.RemoveComponentWithInstanceId0, control.InstanceId), Properties.Resources.Confirm))
                 {
                     control.Delete();