+++ src/Commands/Pages/SavePageConversionLog.cs
@@ -3,11 +3,14 @@ using System;
 using System.IO;
 using System.Management.Automation;
 using System.Reflection;
+using PnP.PowerShell.Commands.Attributes;
 
 namespace PnP.PowerShell.Commands.Pages
 {
     [Cmdlet(VerbsData.Save, "PnPPageConversionLog")]
     [Alias("Save-PnPClientSidePageConversionLog")]
+    [WriteAliasWarning("Please use 'Save-PnPPageConversionLog'. The alias 'Save-PnPClientSidePageConversionLog' will be removed in the 1.5.0 release")]
+
     public class SavePageConversionLog : PnPWebCmdlet
     {
         private Assembly sitesCoreAssembly;