+++ src/Commands/Principals/AddGroupMember.cs
@@ -1,12 +1,13 @@
 ﻿using System.Management.Automation;
 using Microsoft.SharePoint.Client;
-
+using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace PnP.PowerShell.Commands.Principals
 {
     [Cmdlet(VerbsCommon.Add, "PnPGroupMember")]
     [Alias("Add-PnPUserToGroup")]
+    [WriteAliasWarning("Please use Add-PnPGroupMember. The alias `Add-PnPUserToGroup` will be removed in the 1.5.0 release")]
     public class AddUserToGroup : PnPWebCmdlet
     {