+++ src/Commands/Principals/GetGroupMember.cs
@@ -1,13 +1,14 @@
 ﻿using System.Linq;
 using System.Management.Automation;
 using Microsoft.SharePoint.Client;
-
+using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace PnP.PowerShell.Commands.Principals
 {
     [Cmdlet(VerbsCommon.Get, "PnPGroupMember")]
     [Alias("Get-PnPGroupMembers")]
+    [WriteAliasWarning("Please use Get-PnPGroupMember (singular). The alias `Get-PnPGroupMembers` (plural) will be removed in the 1.5.0 release")]
     public class GetGroupMembers : PnPWebCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true)]
@@ -18,7 +19,6 @@ namespace PnP.PowerShell.Commands.Principals
         public string User;
         protected override void ExecuteCmdlet()
         {
-
             if (ParameterSpecified(nameof(User)))
             {
                 var g = Group.GetGroup(PnPContext);