+++ src/Commands/Principals/RemoveGroupMember.cs
@@ -2,12 +2,14 @@
 using System.Management.Automation;
 using Microsoft.SharePoint.Client;
 using PnP.Core;
+using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace PnP.PowerShell.Commands.Principals
 {
     [Cmdlet(VerbsCommon.Remove, "PnPGroupMember")]
     [Alias("Remove-PnPUserFromGroup")]
+    [WriteAliasWarning("Please use Remove-PnPGroupMember. The alias `Remove-PnPUserFromGroup` will be removed in the 1.5.0 release")]
     public class RemoveUserFromGroup : PnPWebCmdlet
     {